Move headers to subdirectory
so they will land in /usr/include/MiniLua instead of the root
/usr/include when installing.
