+++ examples/cli/main.cpp
@@ -1,5 +1,5 @@
-#include "luaparser.h"
-#include "luainterpreter.h"
+#include "MiniLua/luaparser.hpp"
+#include "MiniLua/luainterpreter.hpp"
 #include <chrono>
 #include <vector>
