+++ include/MiniLua/sourcechange.hpp
@@ -1,8 +1,8 @@
 #ifndef SOURCECHANGE_H
 #define SOURCECHANGE_H
 
-#include "val.h"
-#include "luatoken.h"
+#include "val.hpp"
+#include "luatoken.hpp"
 
 namespace lua {
 namespace rt {
