+++ include/MiniLua/sourceexp.hpp
@@ -1,8 +1,8 @@
 #ifndef SOURCEEXP_H
 #define SOURCEEXP_H
 
-#include "val.h"
-#include "luatoken.h"
+#include "val.hpp"
+#include "luatoken.hpp"
 
 namespace lua {
 namespace rt {
