+++ src/core/luatoken.cpp
@@ -1,4 +1,4 @@
-#include "luatoken.h"
+#include "MiniLua/luatoken.hpp"
 
 string LuaToken::to_string() const {
     return "[" + std::to_string(static_cast<int>(type)) + "]" + match + " start:" + std::to_string(pos) + " length:" + std::to_string(length);
